+++ src/defines.h
@@ -13,7 +13,7 @@
 //-------------------------------------
 #define VERSION_MAJOR       0
 #define VERSION_MINOR       8
-#define VERSION_PATCH       173
+#define VERSION_PATCH       178
 //-------------------------------------
 typedef struct {
     uint8_t ch;